Extract a base connection class to for service connection and utilities (#153)
* Extract a base connection class to for basic service connection actions, such as keepalive, handshake, and process incoming messages to a Common csproj
* Change all class to internal as it is a SDK project
