+++ src/Microsoft.Azure.SignalR.Common/IServiceConnection.cs
@@ -1,0 +1,17 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Threading.Tasks;
+using Microsoft.Azure.SignalR.Protocol;
+
+namespace Microsoft.Azure.SignalR
+{
+    internal interface IServiceConnection
+    {
+        Task StartAsync();
+
+        Task WriteAsync(ServiceMessage serviceMessage);
+
+        Task StopAsync();
+    }
+}
