+++ src/Microsoft.Azure.SignalR.Common/StaticRandom.cs
